+++ drivers/include/nrfx_comp.h
@@ -106,9 +106,7 @@ typedef struct
  * This function initializes the COMP driver, but does not enable the peripheral or any interrupts.
  * To start the driver, call the function @ref nrfx_comp_start() after initialization.
  *
- * If no configuration structure is provided, the driver is initialized with the default settings.
- *
- * @param[in] p_config      Initial configuration. If NULL, the default configuration is used.
+ * @param[in] p_config      Pointer to the structure with initial configuration.
  * @param[in] event_handler Handler function.
  *
  * @retval NRFX_ERROR_INVALID_PARAM If the configuration is invalid.