+++ drivers/src/nrfx_qspi.c
@@ -81,6 +81,7 @@ nrfx_err_t nrfx_qspi_init(nrfx_qspi_config_t const * p_config,
                           nrfx_qspi_handler_t        handler,
                           void *                     p_context)
 {
+    NRFX_ASSERT(p_config);
     if (m_cb.state != NRFX_DRV_STATE_UNINITIALIZED)
     {
         return NRFX_ERROR_INVALID_STATE;